+++ examples/simple-generator/src/app/index.ts
@@ -10,7 +10,11 @@ import type { Ui5App } from '@sap-ux/ui5-application-writer';
 import { isAppStudio } from '@sap-ux/btp-utils';
 import type { ServiceInfo } from './service';
 import { getServiceInfo, getServiceInfoInBAS } from './service';
+import { getUI5Versions } from './ui5';
 
+/**
+ *
+ */
 export default class extends Generator {
     private app!: Ui5App & { app: { flpAppId: string } };
     private service!: OdataService;
@@ -56,6 +60,7 @@ export default class extends Generator {
             type: 'input',
             name: 'entity',
             message: 'Main entity',
+            default: 'SEPMRA_C_PD_Product',
             validate: (answer) => !!answer
         });
 
@@ -63,7 +68,8 @@ export default class extends Generator {
             version: OdataVersion.v2,
             url: service.url,
             path: service.path,
-            metadata: service.metadata
+            metadata: service.metadata,
+            annotations: service.annotations
         };
         if (service.destination) {
             this.service.destination = {
@@ -120,7 +126,7 @@ export default class extends Generator {
                     ...this.app,
                     service: this.service,
                     ui5: {
-                        minUI5Version: '1.99.0'
+                        minUI5Version: (await getUI5Versions()).latest.version
                     },
                     template: this.template as FreestyleTemplate
                 },