+++ packages/environment-check/jest.config.js
@@ -1,0 +1,20 @@
+module.exports = {
+    transform: {
+        '^.+\\.ts$': 'ts-jest'
+    },
+    collectCoverage: true,
+    collectCoverageFrom: ['src/**/*.ts'],
+    reporters: [
+        'default',
+        [
+            'jest-sonar',
+            {
+                reportedFilePath: 'relative',
+                relativeRootDir: '<rootDir>/../../../'
+            }
+        ]
+    ],
+    coverageReporters: ['text', ['lcov', { projectRoot: '../../' }]],
+    modulePathIgnorePatterns: ['<rootDir>/dist', '<rootDir>/coverage'],
+    verbose: true
+};