+++ packages/fiori-freestyle-writer/src/index.ts
@@ -8,7 +8,7 @@ import { getPackageJsonTasks } from './packageConfig';
 import cloneDeep from 'lodash/cloneDeep';
 import type { BasicAppSettings } from './types';
 import { FreestyleApp, TemplateType } from './types';
-import { setDefaults } from './defaults';
+import { setDefaults, escapeFLPText } from './defaults';
 
 /**
  * Generate a UI5 application based on the specified Fiori Freestyle floorplan template.
@@ -20,7 +20,7 @@ import { setDefaults } from './defaults';
  */
 async function generate<T>(basePath: string, data: FreestyleApp<T>, fs?: Editor): Promise<Editor> {
     // Clone rather than modifying callers refs
-    const ffApp: FreestyleApp<T> = cloneDeep(data) as FreestyleApp<T>;
+    const ffApp = cloneDeep(data);
     // set defaults
     setDefaults(ffApp);
 
@@ -29,16 +29,19 @@ async function generate<T>(basePath: string, data: FreestyleApp<T>, fs?: Editor)
     // add new and overwrite files from templates e.g.
     const tmplPath = join(__dirname, '..', 'templates');
     // Common files
-    fs.copyTpl(join(tmplPath, 'common', 'add', '**/*.*'), basePath, ffApp);
-
-    fs.copyTpl(join(tmplPath, ffApp.template.type, 'add', `**/*.*`), basePath, ffApp, undefined, {});
+    const ignore = [ffApp.appOptions?.typescript ? '**/*.js' : '**/*.ts'];
+    fs.copyTpl(join(tmplPath, 'common', 'add'), basePath, { ...ffApp, escapeFLPText }, undefined, {
+        globOptions: { ignore }
+    });
+    fs.copyTpl(join(tmplPath, ffApp.template.type, 'add'), basePath, ffApp, undefined, { globOptions: { ignore } });
 
     if (ffApp.template.type === TemplateType.Basic) {
         const viewName = (ffApp.template.settings as BasicAppSettings).viewName;
         const viewTarget = join(basePath, 'webapp', 'view', `${viewName}.view.xml`);
         fs.copyTpl(join(tmplPath, ffApp.template.type, 'custom/View.xml'), viewTarget, ffApp);
-        const controllerTarget = join(basePath, `webapp/controller/${viewName}.controller.js`);
-        fs.copyTpl(join(tmplPath, ffApp.template.type, 'custom/Controller.js'), controllerTarget, ffApp);
+        const ext = ffApp.appOptions?.typescript ? 'ts' : 'js';
+        const controllerTarget = join(basePath, `webapp/controller/${viewName}.controller.${ext}`);
+        fs.copyTpl(join(tmplPath, ffApp.template.type, `custom/Controller.${ext}`), controllerTarget, ffApp);
     }
 
     // Add template specific manifest settings