+++ packages/fiori-freestyle-writer/templates/listdetail/add/webapp/controller/DetailObjectNotFound.js
@@ -1,7 +1,0 @@
-sap.ui.define([
-    "./BaseController"
-], function (BaseController) {
-    "use strict";
-
-    return BaseController.extend("<%- app.id %>.controller.DetailObjectNotFound", {});
-});