+++ packages/fiori-freestyle-writer/templates/worklist/add/webapp/controller/App.controller.ts
@@ -1,0 +1,12 @@
+import BaseController from "./BaseController";
+
+/**
+ * @namespace <%- app.id %>
+ */
+export default class App extends BaseController {
+
+    public onInit(): void {
+        // apply content density mode to root view
+        this.getView()!.addStyleClass(this.getUIComponent().getContentDensityClass());
+    }
+}